+++ src/main/java/persistence/dialect/type/JavaToH2Type.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.dialect.type;
+package persistence.dialect.type;
 
 public enum JavaToH2Type {
     LONG("bigint"),
