+++ src/test/java/persistence/sql/ddl/collection/IdGeneratedValueStrategyMapTest.java
@@ -3,6 +3,7 @@ package persistence.sql.ddl.collection;
 import jakarta.persistence.GenerationType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import persistence.dialect.collection.IdGeneratedValueStrategyMap;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
